chore(filecoin-wallet-provider): test in TS
