+++ packages/filecoin-wallet-provider/tsconfig.json
@@ -3,9 +3,11 @@
     "target": "es5",
     "module": "commonjs",
     "declaration": true,
+    "strict": true,
+    "esModuleInterop": true,
+    "sourceMap": true,
     "outDir": "./dist",
-    "strict": true
+    "sourceRoot": "src/"
   },
-  "include": ["src"],
   "exclude": ["node_modules", "**/__tests__/*"]
 }
